agents/incident-triage/pyproject.toml
[project]
name = "agent-hub-incident-triage"
version = "0.1.0"
description = "Incident triage agent — FastAPI + settings (incremental build)"
requires-python = ">=3.11"
dependencies = [
  "agent-hub-core",
  "asyncpg>=0.29.0",
  "boto3>=1.34.0",
  "fastapi>=0.115.0",
  "google-api-python-client>=2.0",
  "google-auth>=2.0",
  "google-auth-httplib2>=0.2",
  "httpx>=0.27.0",
  "langchain>=0.3",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.3",
  "langfuse>=3.0",
  "langgraph>=0.2.0",
  "langgraph-checkpoint-postgres>=0.2",
  "psycopg[binary]>=3.2",
  "pydantic-settings>=2.6.0",
  "slack-sdk>=3.27",
  "sqlalchemy[asyncio]>=2.0.0",
  "uvicorn[standard]>=0.32.0",
]

[tool.uv.sources]
agent-hub-core = { workspace = true }

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["incident_triage*"]
